[library][cpp] test EContentAdapter with a random library object
